Refactor: [WalletConnect] lazy-load based on feature toggle
